backend/requirements.txt
# FastAPI 后端依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# 阿里云 DashScope SDK
dashscope>=1.25.0

# 文件解析
python-docx>=1.1.0
pypdf>=4.0.0

# HTTP 客户端
aiohttp>=3.9.0
requests>=2.31.0

# 工具库
python-dotenv>=1.0.0
